fix: dto의 평점 타입 BigDecimal로 변경, isbn과 썸네일 url값을 강제하지 않도록 수정

+++ src/main/java/com/codeit/mission/deokhugam/book/dto/BookDto.java
@@ -1,5 +1,6 @@
 package com.codeit.mission.deokhugam.book.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -14,7 +15,7 @@ public record BookDto(
     String isbn,
     String thumbnailUrl,
     long reviewCount,
-    double rating,
+    BigDecimal rating,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
 ) {
